+++ MacPacker/Features/Welcome/WelcomeView.swift
@@ -72,6 +72,7 @@ struct WelcomeWhatsNewView: View {
             WhatsNewPill(key: LocalizedStringResource("v0.15_pr_70", defaultValue: "Support for wim disk images", table: "LocalizableWhatsNew"), type: .feature)
             WhatsNewPill(key: LocalizedStringResource("v0.15_issue_6", defaultValue: "Open password protected / ecrypted archives", table: "LocalizableWhatsNew"), type: .feature)
             WhatsNewPill(key: LocalizedStringResource("v0.15_issue_82", defaultValue: "New About dialog", table: "LocalizableWhatsNew"), type: .feature)
+            WhatsNewPill(key: LocalizedStringResource("v0.15_issue_78", defaultValue: "Up/down keys not working when preview is opened", table: "LocalizableWhatsNew"), type: .bug)
             WhatsNewPill(key: LocalizedStringResource("v0.15_pr_69", defaultValue: "Support for magic numbers at the end of archive files", table: "LocalizableWhatsNew"), type: .core)
             WhatsNewPill(key: LocalizedStringResource("v0.15_pr_79", defaultValue: "Update acknowledgements", table: "LocalizableWhatsNew"), type: .core)
             WhatsNewPill(key: LocalizedStringResource("v0.15_issue_80", defaultValue: "Drop macOS 13 support", table: "LocalizableWhatsNew"), type: .core)